pyproject.toml
[tool.poetry]
name = "fcrypt"
version = "1.0-b1"
description = "Encrypt file line by lines"
authors = ["XyzzyZ4p <xyzzyzap@gmail.com>"]
readme = "README.md"
packages = [{include = "fcrypt"}]

[tool.poetry.dependencies]
python = "^3.11"
pycryptodome = "^3.16.0"
python-dotenv = "^0.21.1"
click = "^8.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
